登录账户id超过Int类型范围（protobuf中需要定义为int64以对应jvm中long类型）问题修复

+++ app/src/main/java/com/mrlin/composemany/repository/entity/User.kt
@@ -22,14 +22,14 @@ data class User(
     val token: String? = null,
     val cookie: String? = null,
     @PrimaryKey
-    var accountId: Int = account.id,
+    var accountId: Long = account.id,
 ) {
     fun isValid() = code < 299
 }
 
 data class Account(
     @ColumnInfo(name = "ua_id")
-    val id: Int,
+    val id: Long,
     @ColumnInfo(name = "ua_user_name")
     val userName: String,
     @ColumnInfo(name = "ua_type")
